test(testing): query и reading tests вынесены в отдельные suite files
query/read-model, reading-shell state и reading-specific presentation/navigation tests выделены из RSSReaderTests.swift в самостоятельные feature-oriented suites
