+++ RSSReaderTests/Feeds/FeedNormalizationTests.swift
@@ -1,0 +1,75 @@
+import Foundation
+import Testing
+@testable import RSSReader
+
+@Suite("Feeds / Normalization")
+@MainActor
+struct FeedNormalizationTests {
+    @Test
+    func feedNormalizationKeepsFaviconLikeIconURLAndNormalizesIt() {
+        let feed = ParsedFeedDTO(
+            kind: .rss,
+            metadata: ParsedFeedMetadataDTO(
+                title: "Example Feed",
+                siteURL: "HTTPS://Example.com",
+                iconURL: "HTTPS://CDN.EXAMPLE.COM/Favicon-32x32.png?cache=1#fragment"
+            ),
+            entries: []
+        )
+
+        let normalized = FeedNormalizationService.normalize(feed, feedURL: "https://example.com/feed.xml")
+
+        #expect(normalized.metadata.siteURL == "https://example.com/")
+        #expect(normalized.metadata.iconURL == "https://cdn.example.com/Favicon-32x32.png?cache=1")
+    }
+
+    @Test
+    func feedNormalizationRewritesLogoAssetToSiteFaviconWhenSiteURLIsKnown() {
+        let feed = ParsedFeedDTO(
+            kind: .rss,
+            metadata: ParsedFeedMetadataDTO(
+                title: "Example Feed",
+                siteURL: "https://example.com/news/",
+                iconURL: "https://cdn.example.com/assets/header-logo.png"
+            ),
+            entries: []
+        )
+
+        let normalized = FeedNormalizationService.normalize(feed, feedURL: "https://example.com/feed.xml")
+
+        #expect(normalized.metadata.iconURL == "https://example.com/favicon.ico")
+    }
+
+    @Test
+    func feedNormalizationKeepsOriginalIconURLWhenItCannotBuildSiteFaviconFallback() {
+        let feed = ParsedFeedDTO(
+            kind: .atom,
+            metadata: ParsedFeedMetadataDTO(
+                title: "Example Feed",
+                iconURL: "https://cdn.example.com/assets/banner-logo.png"
+            ),
+            entries: []
+        )
+
+        let normalized = FeedNormalizationService.normalize(feed, feedURL: "https://example.com/feed.xml")
+
+        #expect(normalized.metadata.iconURL == "https://cdn.example.com/assets/banner-logo.png")
+    }
+
+    @Test
+    func feedNormalizationUsesSiteFaviconWhenFeedDidNotProvideIconURL() {
+        let feed = ParsedFeedDTO(
+            kind: .rss,
+            metadata: ParsedFeedMetadataDTO(
+                title: "Example Feed",
+                siteURL: "HTTPS://Example.com/news"
+            ),
+            entries: []
+        )
+
+        let normalized = FeedNormalizationService.normalize(feed, feedURL: "https://example.com/feed.xml")
+
+        #expect(normalized.metadata.siteURL == "https://example.com/news")
+        #expect(normalized.metadata.iconURL == "https://example.com/favicon.ico")
+    }
+}
