+++ RSSReaderTests/Feeds/SourceIconCacheServiceTests.swift
@@ -1,0 +1,57 @@
+import Foundation
+import Testing
+@testable import RSSReader
+
+@Suite("Feeds / Source Icon Cache")
+@MainActor
+struct SourceIconCacheServiceTests {
+    @Test
+    func sourceIconCacheReturnsCachedDataWithoutSecondNetworkRequest() async throws {
+        let iconURL = try #require(URL(string: "https://example.com/favicon.ico"))
+        let httpClient = ScriptedHTTPClient(
+            responsesByURL: [
+                iconURL.absoluteString: .response(
+                    statusCode: 200,
+                    headers: ["Content-Type": "image/x-icon"],
+                    body: "icon-binary"
+                )
+            ]
+        )
+        let service = SourceIconCacheService(httpClient: httpClient)
+
+        let firstLoad = try await service.imageData(for: iconURL)
+        let secondLoad = try await service.imageData(for: iconURL)
+
+        #expect(firstLoad == Data("icon-binary".utf8))
+        #expect(secondLoad == firstLoad)
+
+        let requests = await httpClient.recordedRequests()
+        #expect(requests.count == 1)
+    }
+
+    @Test
+    func sourceIconCacheSharesInFlightRequestBetweenConcurrentConsumers() async throws {
+        let iconURL = try #require(URL(string: "https://example.com/favicon.ico"))
+        let httpClient = ScriptedHTTPClient(
+            responsesByURL: [
+                iconURL.absoluteString: .delayedResponse(
+                    statusCode: 200,
+                    headers: ["Content-Type": "image/x-icon"],
+                    body: "icon-binary",
+                    delayNanoseconds: 50_000_000
+                )
+            ]
+        )
+        let service = SourceIconCacheService(httpClient: httpClient)
+
+        async let firstLoad = service.imageData(for: iconURL)
+        async let secondLoad = service.imageData(for: iconURL)
+        let (firstResult, secondResult) = try await (firstLoad, secondLoad)
+
+        #expect(firstResult == secondResult)
+
+        let requests = await httpClient.recordedRequests()
+        #expect(requests.count == 1)
+        #expect(await httpClient.maxConcurrentExecutions() == 1)
+    }
+}
