+++ RSSReaderTests/Shell/ShellModalPresentationTests.swift
@@ -1,0 +1,67 @@
+import Foundation
+import Testing
+@testable import RSSReader
+
+@Suite("Shell / Modal Presentation")
+@MainActor
+struct ShellModalPresentationTests {
+    @Test
+    func shellActionEntryPointsUpdateSelectionAndFilterInAppState() throws {
+        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
+        let appState = AppState()
+        let feedID = UUID()
+        let articleID = UUID()
+
+        harness.dependencies.showFeed(id: feedID, using: appState)
+        harness.dependencies.selectArticle(id: articleID, using: appState)
+        harness.dependencies.applySourcesFilter(.unread, using: appState)
+
+        #expect(appState.selectedSidebarSelection == .feed(feedID))
+        #expect(appState.selectedArticleID == articleID)
+        #expect(appState.selectedDetailRoute == .article(articleID))
+        #expect(appState.selectedSourcesFilter == .unread)
+
+        harness.dependencies.showInbox(using: appState)
+
+        #expect(appState.selectedSidebarSelection == .inbox)
+        #expect(appState.selectedArticleID == nil)
+        #expect(appState.selectedDetailRoute == .none)
+
+        harness.dependencies.showFolder(named: "News", using: appState)
+
+        #expect(appState.selectedSidebarSelection == .folder("News"))
+        #expect(appState.selectedArticleID == nil)
+        #expect(appState.selectedDetailRoute == .none)
+
+        harness.dependencies.showUnread(using: appState)
+        #expect(appState.selectedSidebarSelection == .unread)
+
+        harness.dependencies.showStarred(using: appState)
+        #expect(appState.selectedSidebarSelection == .starred)
+    }
+
+    @Test
+    func settingsPresentationStateLivesInAppStateAndDoesNotResetReadingShellContext() throws {
+        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
+        let appState = AppState()
+        let feedID = UUID()
+        let articleID = UUID()
+
+        harness.dependencies.showFeed(id: feedID, using: appState)
+        harness.dependencies.selectArticle(id: articleID, using: appState)
+
+        harness.dependencies.showSettings(using: appState)
+
+        #expect(appState.isPresentingSettingsScreen)
+        #expect(appState.selectedSidebarSelection == .feed(feedID))
+        #expect(appState.selectedArticleID == articleID)
+        #expect(appState.selectedDetailRoute == .article(articleID))
+
+        harness.dependencies.dismissSettings(using: appState)
+
+        #expect(appState.isPresentingSettingsScreen == false)
+        #expect(appState.selectedSidebarSelection == .feed(feedID))
+        #expect(appState.selectedArticleID == articleID)
+        #expect(appState.selectedDetailRoute == .article(articleID))
+    }
+}
