+++ RSSReaderTests/Sidebar/SidebarFilteringTests.swift
@@ -1,0 +1,129 @@
+import Foundation
+import Testing
+@testable import RSSReader
+
+@Suite("Sidebar / Filtering")
+@MainActor
+struct SidebarFilteringTests {
+    @Test
+    func sourcesFilterArticleListFilterResolverMapsSourcesFilterToExpectedArticleFilter() {
+        #expect(SourcesFilterArticleListFilterResolver.resolve(for: .allItems) == .all)
+        #expect(SourcesFilterArticleListFilterResolver.resolve(for: .unread) == .unread)
+        #expect(SourcesFilterArticleListFilterResolver.resolve(for: .starred) == .starred)
+    }
+
+    @Test
+    func sourcesSmartViewsShowOnlyActiveFilterRow() {
+        #expect(SmartSidebarItem.visibleItems(for: .allItems, hasFeeds: true) == [.allItems])
+        #expect(SmartSidebarItem.visibleItems(for: .unread, hasFeeds: true) == [.unread])
+        #expect(SmartSidebarItem.visibleItems(for: .starred, hasFeeds: true) == [.starred])
+    }
+
+    @Test
+    func sourcesSmartViewsAreHiddenWhenThereAreNoFeeds() {
+        #expect(SmartSidebarItem.visibleItems(for: .allItems, hasFeeds: false).isEmpty)
+        #expect(SmartSidebarItem.visibleItems(for: .unread, hasFeeds: false).isEmpty)
+        #expect(SmartSidebarItem.visibleItems(for: .starred, hasFeeds: false).isEmpty)
+    }
+
+    @Test
+    func sourcesSidebarShowsOnlyFeedsWithStarredArticlesWhenStarredFilterIsActive() {
+        let feedOneID = UUID()
+        let feedTwoID = UUID()
+        let newsFolder = Folder(name: "News")
+        let feeds = [
+            FeedSidebarItem(
+                feed: Feed(id: feedOneID, url: "https://example.com/feed-one.xml", title: "Feed One", folder: newsFolder),
+                unreadCount: 2
+            ),
+            FeedSidebarItem(
+                feed: Feed(id: feedTwoID, url: "https://example.com/feed-two.xml", title: "Feed Two"),
+                unreadCount: 0
+            )
+        ]
+
+        let filteredFeeds = SidebarFeedVisibility.filteredFeeds(
+            feeds: feeds,
+            filter: .starred,
+            starredFeedIDs: [feedTwoID]
+        )
+
+        #expect(filteredFeeds.map(\.id) == [feedTwoID])
+    }
+
+    @Test
+    func sourcesSidebarKeepsAllFeedsVisibleForAllItemsFilter() {
+        let feedOneID = UUID()
+        let feedTwoID = UUID()
+        let newsFolder = Folder(name: "News")
+        let feeds = [
+            FeedSidebarItem(
+                feed: Feed(id: feedOneID, url: "https://example.com/feed-one.xml", title: "Feed One", folder: newsFolder),
+                unreadCount: 2
+            ),
+            FeedSidebarItem(
+                feed: Feed(id: feedTwoID, url: "https://example.com/feed-two.xml", title: "Feed Two"),
+                unreadCount: 0
+            )
+        ]
+
+        let allItemsFeeds = SidebarFeedVisibility.filteredFeeds(
+            feeds: feeds,
+            filter: .allItems,
+            starredFeedIDs: [feedTwoID]
+        )
+
+        #expect(allItemsFeeds.map(\.id) == feeds.map(\.id))
+        #expect(allItemsFeeds.map(\.unreadCount) == feeds.map(\.unreadCount))
+    }
+
+    @Test
+    func sourcesSidebarShowsOnlyFeedsWithUnreadArticlesWhenUnreadFilterIsActive() {
+        let feedOneID = UUID()
+        let feedTwoID = UUID()
+        let newsFolder = Folder(name: "News")
+        let feeds = [
+            FeedSidebarItem(
+                feed: Feed(id: feedOneID, url: "https://example.com/feed-one.xml", title: "Feed One", folder: newsFolder),
+                unreadCount: 2
+            ),
+            FeedSidebarItem(
+                feed: Feed(id: feedTwoID, url: "https://example.com/feed-two.xml", title: "Feed Two"),
+                unreadCount: 0
+            )
+        ]
+
+        let filteredFeeds = SidebarFeedVisibility.filteredFeeds(
+            feeds: feeds,
+            filter: .unread,
+            starredFeedIDs: []
+        )
+
+        #expect(filteredFeeds.map(\.id) == [feedOneID])
+    }
+
+    @Test
+    func sourcesSidebarHidesFoldersSectionWhenFilteredFeedsDoNotContainFolders() {
+        let ungroupedFeed = FeedSidebarItem(
+            feed: Feed(id: UUID(), url: "https://example.com/feed.xml", title: "Ungrouped Feed"),
+            unreadCount: 1
+        )
+
+        let groups = FolderSidebarGroup.groups(from: [ungroupedFeed])
+
+        #expect(groups.isEmpty)
+    }
+
+    @Test
+    func sourcesSidebarHidesUngroupedSectionWhenFilteredFeedsDoNotContainUngroupedSources() {
+        let folder = Folder(name: "News")
+        let groupedFeed = FeedSidebarItem(
+            feed: Feed(id: UUID(), url: "https://example.com/feed.xml", title: "Grouped Feed", folder: folder),
+            unreadCount: 1
+        )
+
+        let ungroupedFeeds = SidebarUngroupedFeeds.visibleFeeds(from: [groupedFeed])
+
+        #expect(ungroupedFeeds.isEmpty)
+    }
+}
